using bigdecimal instead of float to store longitude and latitude so that result won't be truncated

+++ src/main/java/com/example/demo/RequestForm/RegisterUserLocationReqForm.java
@@ -1,9 +1,11 @@
 package com.example.demo.RequestForm;
 
+import java.math.BigDecimal;
+
 public class RegisterUserLocationReqForm {
     public String firebaseToken;
 
-    public float longitude;
+    public BigDecimal longitude;
 
-    public float latitude;
+    public BigDecimal latitude;
 }
